moldouber/api-gateway/requirements.txt
Flask==2.2.2
grpcio>=1.66.2
protobuf>=5.28.2
Werkzeug==2.2.3
